requirements/local.txt
asgiref==3.5.0
Django==4.0.2
django-ckeditor==6.2.0
django-js-asset==2.0.0
django-model-utils==4.2.0
mysqlclient==2.1.0
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
Unipath==1.1
autopep8==1.5.7
pycodestyle==2.7.0
toml==0.10.2